ia/requirements.txt
Flask
Flask-CORS
google-genai
gunicorn app:app